feat: Implementar selección de tipo de formulario para preguntas y mejorar la gestión de campos
- Se añadió un modal para seleccionar el tipo de formulario (personalizado o solicitud de acceso) en el componente de configuración de preguntas.
- Se integró el manejo de tipos en el backend para guardar y recuperar los tipos de formulario de las preguntas.
- Se mejoró la gestión de campos con una interfaz de usuario mejorada para agregar, editar y eliminar campos.
- Se implementaron modales de confirmación para acciones de eliminación de campos.
- Se actualizaron los tipos de campos con descripciones para una mejor guía al usuario.
- Se añadió funcionalidad para filtrar preguntas según los permisos del usuario.
- Se creó una migración para agregar la columna 'tipo' a la tabla 'formulario_templates'.

+++ backend/database/migrations/2026_04_18_000001_create_notificaciones_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->enum('tipo', ['voto_util', 'voto_no_util']);
-            $table->unsignedInteger('pregunta_id');
-            $table->integer('modulo_id')->nullable();
+            $table->unsignedBigInteger('pregunta_id');
+            $table->unsignedBigInteger('modulo_id')->nullable();
             $table->string('mensaje');
             $table->boolean('leida')->default(false);
             $table->timestamps();
